Allow l2met-shuttle to operate like `tee`.
This will help improve some signal handling in start-l2met-shuttle (a wrapper
which attaches l2met-shuttle to stdout and stderr of a process). Right now,
that wrapper operates by using `tee` to forward streams to a fifo connected to
l2met-shuttle. If l2met-shuttle dies, the fifo simply goes unread, and tee
eventually blocks.

This change will allow the wrapper to connect stdout and stderr directly to
l2met-shuttle, so that if it exits the upstream process will receive SIGPIPE
and exit.

For compatibility with existing applications, this is optional, rather than the
new default behavior.